Add simpler example to fix example app on pub.dev (#20)
* Remove unneeded imports

* Update Xcode files

* Rename folder and example/lib/main.dart to complex_example_main.dart

* Add very simple example app with link to complex example app

* Refactor folder structure, update link

* Small refactoring

* Update README.md

* Remove docs comment

* Update changelog and version

* Update CHANGELOG.md


* Update README.md


* Remove theme argument

* Add ValueNotifier and ValueListenableBuilder to example app
